Creating Jsonnet implementation
This is a major change in how this works, and should aid in long term
maintainability. JSON is hard to parse for a human in addition to having
a LOT of whitespace, braces, and brackets littered all over the place.

With this change, the goal is both to reduce the cognitive overhead and
to make changes easier. E.g. to add a new terminal emulator to ignore,
it only has to be added in one spot instead of ~50.